requirements.txt
locust
psycopg2-binary
boto3
streamlit
pandas